requirements.txt
fastapi==0.118.0
uvicorn==0.24.0
pydantic==2.9.0
pytest==8.4.3
python-dateutil==2.8.2
psycopg2-binary==2.9.10
python-dotenv==1.0.0
